# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Zaimik API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/",
    "paths": {
        "/": {
            "post": {
                "description": "api endpoint that accepts the user's mail along with the authorization code, checks the correctness of the code and sets cookies (session_id) in the browser. responds with status 200 if successful, with status 400 if code is invalid or email is invalid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publicAuth"
                ],
                "summary": "signIn",
                "parameters": [
                    {
                        "description": "email and code",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.signInReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/catalog": {
            "put": {
                "description": "unfinished, because the online cash register is not connected, do not use this",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "unfinished"
                ],
                "summary": "updateUserGetLinkOnPayment",
                "parameters": [
                    {
                        "description": "user update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.suitableCompanies"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.suitableCompanies"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "when user successfully pay for subscription, he can get sorted catalog of loan companies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "getSortedSuitableCatalog",
                "parameters": [
                    {
                        "description": "user update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.suitableCompanies"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.suitableCompanies"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/check-subscription": {
            "get": {
                "description": "after payment for subscription endpoint can answer what status of subscription is it now",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "checkSubscription",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.checkSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/send-review": {
            "post": {
                "description": "api endpoint, accepts feedback from a person if he is authorized, if not authorized, responds with an error",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "sendReview",
                "parameters": [
                    {
                        "description": "info of review",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.sendReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/sign-out": {
            "post": {
                "description": "api endpoint, logs out the user if he is logged in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "signOut",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/get-code": {
            "post": {
                "description": "api endpoint that receives the user's mail, an authorization code is sent to it, which must be entered in a special field, the code lives for 1 minute. responds with status 400 if invalid data is sent or invalid mail is sent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "publicAuth"
                ],
                "summary": "getAuthCode",
                "parameters": [
                    {
                        "description": "email",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.getCodeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/reviews": {
            "get": {
                "description": "api endpoint, send all moderated reviews which consist in dataBase in json format",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "getReviews",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.reviewsResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.checkSubscriptionResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "expired_date": {
                    "type": "string"
                }
            }
        },
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "handler.getCodeReq": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "handler.reviewsResponse": {
            "type": "object",
            "properties": {
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Review"
                    }
                }
            }
        },
        "handler.sendReviewRequest": {
            "type": "object",
            "required": [
                "review"
            ],
            "properties": {
                "review": {
                    "$ref": "#/definitions/models.Review"
                }
            }
        },
        "handler.signInReq": {
            "type": "object",
            "required": [
                "code",
                "email"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "handler.suitableCompanies": {
            "type": "object",
            "properties": {
                "companies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LoanCompany"
                    }
                }
            }
        },
        "handler.updateReq": {
            "type": "object",
            "properties": {
                "loan": {
                    "$ref": "#/definitions/models.Loan"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "models.Loan": {
            "type": "object",
            "required": [
                "loan_amount",
                "loan_duration"
            ],
            "properties": {
                "loan_amount": {
                    "type": "integer"
                },
                "loan_duration": {
                    "type": "integer"
                }
            }
        },
        "models.LoanCompany": {
            "type": "object",
            "required": [
                "_",
                "link_on_company_site",
                "logo",
                "max_loan_amount",
                "max_loan_duration",
                "min_loan_percent",
                "name",
                "priority"
            ],
            "properties": {
                "_": {
                    "type": "string"
                },
                "link_on_company_site": {
                    "type": "string"
                },
                "logo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "max_loan_amount": {
                    "type": "integer"
                },
                "max_loan_duration": {
                    "type": "integer"
                },
                "min_loan_percent": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                }
            }
        },
        "models.Review": {
            "type": "object",
            "required": [
                "review",
                "reviewer_name",
                "reviwer_phone"
            ],
            "properties": {
                "review": {
                    "type": "string"
                },
                "reviewer_name": {
                    "type": "string"
                },
                "reviwer_phone": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "city",
                "email",
                "first_name",
                "last_name",
                "phone_number",
                "sur_name"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "sur_name": {
                    "type": "string"
                }
            }
        }
    }
}